refactor(onboarding): simplify asset and security setup flows
- Add new asset type selection and deployment components
- Improve wallet security setup with better UX
- Update all onboarding routes to use new components
- Add user metadata fields for onboarding completion tracking
- Simplify welcome page and remove unnecessary complexity
