+++ kit/dapp/src/components/onboarding/assets/asset-type-card.tsx
@@ -1,9 +1,4 @@
 import { Checkbox } from "@/components/ui/checkbox";
-import {
-  FormControl,
-  FormItem,
-  FormLabel,
-} from "@/components/ui/tanstack-form";
 import type { AssetType } from "@/lib/zod/validators/asset-types";
 import type { LucideIcon } from "lucide-react";
 import { memo } from "react";
@@ -40,7 +35,9 @@ export const AssetTypeCard = memo(
 
     // Event handlers are automatically optimized by React Compiler - no manual memoization needed
     const handleItemClick = () => {
-      if (isDisabled) return;
+      if (isDisabled) {
+        return;
+      }
       onToggle(!isChecked);
     };
 
@@ -49,12 +46,14 @@ export const AssetTypeCard = memo(
     };
 
     const handleCheckboxChange = (checked: boolean) => {
-      if (isDisabled) return;
+      if (isDisabled) {
+        return;
+      }
       onToggle(checked);
     };
 
     return (
-      <FormItem
+      <div
         className={`flex flex-row items-start space-x-3 space-y-0 rounded-lg border p-4 transition-colors ${
           isDisabled
             ? "opacity-50 cursor-not-allowed bg-muted/50"
@@ -62,25 +61,23 @@ export const AssetTypeCard = memo(
         }`}
         onClick={handleItemClick}
       >
-        <FormControl>
-          <CheckboxWrapper onClick={handleCheckboxClick}>
-            <Checkbox
-              checked={isChecked}
-              onCheckedChange={handleCheckboxChange}
-              disabled={isDisabled}
-            />
-          </CheckboxWrapper>
-        </FormControl>
+        <CheckboxWrapper onClick={handleCheckboxClick}>
+          <Checkbox
+            checked={isChecked}
+            onCheckedChange={handleCheckboxChange}
+            disabled={isDisabled}
+          />
+        </CheckboxWrapper>
         <div className="flex-1 space-y-1 leading-none">
           <div className="flex items-center gap-2">
             <Icon
               className={`h-4 w-4 ${isDisabled ? "text-muted-foreground/50" : "text-muted-foreground"}`}
             />
-            <FormLabel
+            <label
               className={`text-sm font-medium ${isDisabled ? "" : "cursor-pointer"}`}
             >
               {t(`asset-types.${assetType}`, { ns: "tokens" })}
-            </FormLabel>
+            </label>
             {isDisabled && (
               <span className="text-xs text-muted-foreground">(Deployed)</span>
             )}
@@ -89,7 +86,7 @@ export const AssetTypeCard = memo(
             {t(`assets.descriptions.${assetType}`)}
           </p>
         </div>
-      </FormItem>
+      </div>
     );
   }
 );
